pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["marek.tremel <tremelmarek@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
pydantic-settings = "^2.3.1"
sqlalchemy = "^2.0.30"
pymysql = "^1.1.1"
pytz = "^2024.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
